+++ MaterialLib/MPL/Properties/CreateProperties.h
@@ -26,11 +26,11 @@
 #include "CreateParameterProperty.h"
 #include "CreatePermeabilityOrthotropicPowerLaw.h"
 #include "CreatePorosityFromMassBalance.h"
-#include "CreateRelPermBrooksCorey.h"
-#include "CreateRelPermLiakopoulos.h"
-#include "CreateRelPermVanGenuchten.h"
 #include "CapillaryPressureSaturation/CreateSaturationBrooksCorey.h"
 #include "CreateSaturationDependentSwelling.h"
 #include "CapillaryPressureSaturation/CreateSaturationLiakopoulos.h"
 #include "CapillaryPressureSaturation/CreateSaturationVanGenuchten.h"
 #include "CreateTransportPorosityFromMassBalance.h"
+#include "RelativePermeability/CreateRelPermBrooksCorey.h"
+#include "RelativePermeability/CreateRelPermLiakopoulos.h"
+#include "RelativePermeability/CreateRelPermVanGenuchten.h"
